# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Mercadeiro API
  description: >
    API do Mercadeiro — app de lista de compras domiciliar (Despensa → Carrinho →
    Finalizar Compra → Analítico). Documenta os endpoints da Fase 1 (API core).
    Regras de negócio referenciadas (RN-xx) e casos de uso (UC-xx) estão detalhados
    em `mercadeiro-regras-de-negocio.md`, na raiz do monorepo.


    **Autenticação:** todas as rotas exigem `Authorization: Bearer <token>`,
    exceto `/health`. Token único de aplicação (app single-user, sem contas —
    ver decisão em `mercadeiro-regras-de-negocio.md` §5.8), configurado em
    `API_BEARER_TOKEN` no `.env` do backend. Requisições sem token válido
    recebem `401 Unauthorized`.
  version: 0.1.0
servers:
  - url: http://localhost:3333
    description: Ambiente local de desenvolvimento
security:
  - bearerAuth: []
...
paths:
...
  /supermercados:
    get:
      tags: [Supermercados]
      summary: Lista supermercados
      operationId: listarSupermercados
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Lista de supermercados
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Supermercado"
    post:
      tags: [Supermercados]
      summary: Cria um supermercado, ou reaproveita um existente com mesmo nome+endereço
      description: >
        Usado pelo fluxo de reverse geocoding (UC-03) para não criar um
        Supermercado duplicado a cada compra no mesmo local.
      operationId: criarSupermercado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [nome]
              properties:
                nome:
                  type: string
                  minLength: 1
                endereco:
                  type: string
                latitude:
                  type: number
                longitude:
                  type: number
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Supermercado existente reaproveitado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supermercado"
        "201":
          description: Supermercado criado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supermercado"
        "400":
          $ref: "#/components/responses/BadRequest"
  /supermercados/{id}:
    parameters:
      - $ref: "#/components/parameters/SupermercadoId"
    get:
      tags: [Supermercados]
      summary: Busca um supermercado por id
      operationId: buscarSupermercado
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Supermercado encontrado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supermercado"
        "404":
          $ref: "#/components/responses/NotFound"
    put:
      tags: [Supermercados]
      summary: Atualiza dados de um supermercado
      operationId: atualizarSupermercado
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                nome:
                  type: string
                  minLength: 1
                endereco:
                  type: string
                  nullable: true
                latitude:
                  type: number
                  nullable: true
                longitude:
                  type: number
                  nullable: true
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Supermercado atualizado
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Supermercado"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags: [Supermercados]
      summary: Remove um supermercado sem compras/histórico vinculados
      operationId: removerSupermercado
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "204":
          description: Supermercado removido
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          description: Supermercado possui compras/histórico vinculados
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /supermercados/{id}/historico:
    parameters:
      - $ref: "#/components/parameters/SupermercadoId"
    get:
      tags: [Supermercados]
      summary: Histórico de preço daquele supermercado, agrupado por produto
      description: Base do ranking por supermercado (3b) e do comparativo entre mercados (3c).
      operationId: buscarHistoricoDoSupermercado
      responses:
        "401":
          $ref: "#/components/responses/Unauthorized"
        "200":
          description: Histórico agregado por produto
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/HistoricoProdutoNoSupermercado"
        "404":
          $ref: "#/components/responses/NotFound"
...
components:
  parameters:
...
    SupermercadoId:
      name: id
      in: path
      required: true
      schema:
        type: string
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Token único de aplicação (API_BEARER_TOKEN no .env do backend).
        Exigido em todas as rotas, exceto /health.
  responses:
...
    Unauthorized:
      description: Token de autenticação ausente ou inválido
      headers:
        WWW-Authenticate:
          schema:
            type: string
            example: Bearer
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: Dados inválidos
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: Recurso não encontrado
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        details:
          description: >
            Detalhes adicionais, quando aplicável — formato varia por origem:
            array de issues do Zod em erros de validação, objeto de metadados
            do Prisma em conflitos de unicidade, ou ausente.
...
    Produto:
      type: object
      required: [id, nome, categoriaId, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        codigoBarras:
          type: string
          nullable: true
        categoriaId:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    Supermercado:
      type: object
      required: [id, nome, createdAt, updatedAt]
      properties:
        id:
          type: string
        nome:
          type: string
        endereco:
          type: string
          nullable: true
        latitude:
          type: number
          nullable: true
        longitude:
          type: number
          nullable: true
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    HistoricoProdutoNoSupermercado:
      type: object
      required: [produto, quantidadeRegistros, precoMedio, menorPreco, ultimoPreco]
      properties:
        produto:
          $ref: "#/components/schemas/Produto"
        quantidadeRegistros:
          type: integer
        precoMedio:
          type: number
        menorPreco:
          type: number
        ultimoPreco:
          type: number
...